requirements.txt
PyPDF2
langchain
langchain_google_genai
langchain_community
pinecone
google.generativeai
fpdf